refactoring package name processingUsingCommand to command #9
