+++ src/main/java/sudoku/MainCommandPattern.java
@@ -5,8 +5,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sudoku.exceptions.IllegalSudokuStateException;
 import sudoku.model.Sudoku;
-import sudoku.processingUsingCommand.AutomatedInvoker;
-import sudoku.processingUsingCommand.Command;
+import sudoku.command.AutomatedInvoker;
+import sudoku.command.Command;
 import sudoku.processingUsingStrategy.BacktrackLucia;
 import sudoku.processingUsingStrategy.HiddenSingleInACell;
 import sudoku.processingUsingStrategy.NakedSingleInACell;
