+++ src/main/java/sudoku/command/ManualInvoker.java
@@ -1,4 +1,4 @@
-package sudoku.processingUsingCommand;
+package sudoku.command;
 
 import sudoku.model.Sudoku;
 import sudoku.processingUsingStrategy.NakedSingleInACell;
