+++ src/test/java/sudoku/processingUsingStrategy/invoker/ManualInvokerTest.java
@@ -5,9 +5,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sudoku.exceptions.IllegalSudokuStateException;
 import sudoku.model.Sudoku;
-import sudoku.processingUsingCommand.Command;
-import sudoku.processingUsingCommand.CommandPicker;
-import sudoku.processingUsingCommand.ManualInvoker;
+import sudoku.command.Command;
+import sudoku.command.CommandPicker;
+import sudoku.command.ManualInvoker;
 import sudoku.processingUsingStrategy.BacktrackLuciaTest;
 import sudoku.processingUsingStrategy.NakedSingleInACell;
 import sudoku.readers.FileSudokuReader;
